Closes #21

+++ src/main/java/com/ardecs/ctshop/service/OrderService.java
@@ -1,25 +1,23 @@
 package com.ardecs.ctshop.service;
 
-import com.ardecs.ctshop.exceptions.NotFoundException;
+import com.ardecs.ctshop.exceptions.PaidOrderCannotBeDeletedException;
 import com.ardecs.ctshop.persistence.entity.Order;
-import com.ardecs.ctshop.persistence.entity.OrderProduct;
 import com.ardecs.ctshop.persistence.entity.Product;
 import com.ardecs.ctshop.persistence.repository.OrderRepository;
-import com.ardecs.ctshop.persistence.repository.ProductRepository;
 import org.springframework.stereotype.Service;
 
 import java.math.BigDecimal;
+import java.util.ArrayList;
 import java.util.HashMap;
+import java.util.List;
 import java.util.Map;
 
 @Service
 public class OrderService {
 
-    private final ProductRepository productRepository;
     private final OrderRepository orderRepository;
 
-    public OrderService(ProductRepository productRepository, OrderRepository orderRepository) {
-        this.productRepository = productRepository;
+    public OrderService(OrderRepository orderRepository) {
         this.orderRepository = orderRepository;
     }
 
@@ -30,25 +28,21 @@ public class OrderService {
     }
 
     public BigDecimal getTotalSum(Order order) {
-        BigDecimal totalSum = BigDecimal.valueOf(0);
-        for (OrderProduct orderProduct : order.getOrderProducts()) {
-            totalSum = new BigDecimal
-                    (String.valueOf
-                            (totalSum.add(BigDecimal.valueOf(orderProduct.getProduct().getPrice())
-                                    .multiply(BigDecimal.valueOf(orderProduct.getQuantityInOrder())))));
-        }
-        return totalSum;
-
+        List<BigDecimal> sums = new ArrayList<>();
+        order.getOrderProducts().forEach(p -> sums.add(p.getProduct().getPrice().multiply(BigDecimal.valueOf(p.getQuantityInOrder()))));
+        return sums.stream().reduce(BigDecimal.ZERO, BigDecimal::add);
     }
 
     public void deleteOrder(Order order) {
 
-        for (OrderProduct orderProduct : order.getOrderProducts()) {
-            Product product = productRepository.findById(orderProduct.getProduct().getId())
-                    .orElseThrow(NotFoundException::new);
-            product.setQuantity(product.getQuantity() + orderProduct.getQuantityInOrder());
-            productRepository.save(product);
+        if (order.getIsPaid()) {
+            throw new PaidOrderCannotBeDeletedException("Paid order cannot be deleted");
         }
+
+        order.getOrderProducts().forEach(
+                p -> p.getProduct().setQuantity(p.getProduct().getQuantity() + p.getQuantityInOrder())
+        );
+
         orderRepository.delete(order);
     }
